backend/requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index requirements/base.in
#
amqp==5.0.3
    # via kombu
attrs==20.3.0
    # via jsonschema
billiard==3.6.3.0
    # via celery
celery==5.0.5
    # via django-camunda
certifi==2020.12.5
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
cffi==1.14.4
    # via cryptography
chardet==4.0.0
    # via requests
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.1.6
    # via celery
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
cryptography==3.3.1
    # via django-auth-adfs
defusedxml==0.6.0
    # via django-camunda
django-admin-index==1.4.0
    # via -r requirements/base.in
django-auth-adfs-db==0.2.0
    # via -r requirements/base.in
django-auth-adfs==1.5.0
    # via django-auth-adfs-db
django-axes==5.12.0
    # via -r requirements/base.in
django-camunda==0.10.0
    # via -r requirements/base.in
django-choices==1.7.1
    # via
    #   -r requirements/base.in
    #   zgw-consumers
django-compat==1.0.15
    # via
    #   django-hijack
    #   django-hijack-admin
django-hijack-admin==2.1.10
    # via -r requirements/base.in
django-hijack==2.1.10
    # via
    #   -r requirements/base.in
    #   django-hijack-admin
django-ipware==3.0.2
    # via django-axes
django-ordered-model==3.4.1
    # via django-admin-index
django-privates==1.2.1
    # via -r requirements/base.in
django-redis==4.12.1
    # via -r requirements/base.in
django-relativedelta==1.1.2
    # via zgw-consumers
django-sendfile2==0.6.0
    # via django-privates
django-solo==1.1.5
    # via
    #   django-auth-adfs-db
    #   django-camunda
    #   zgw-consumers
django==2.2.17
    # via
    #   -r requirements/base.in
    #   django-auth-adfs
    #   django-auth-adfs-db
    #   django-axes
    #   django-camunda
    #   django-choices
    #   django-privates
    #   django-redis
    #   django-relativedelta
    #   django-sendfile2
    #   djangorestframework
    #   drf-spectacular
    #   zgw-consumers
djangorestframework-camel-case==1.2.0
    # via -r requirements/base.in
djangorestframework==3.12.2
    # via
    #   -r requirements/base.in
    #   drf-spectacular
drf-spectacular==0.13.1
    # via -r requirements/base.in
elastic-apm==5.10.1
    # via -r requirements/base.in
faker==5.7.0
    # via zgw-consumers
gemma-zds-client==0.14.0
    # via zgw-consumers
idna==2.10
    # via requests
inflection==0.5.1
    # via
    #   django-camunda
    #   drf-spectacular
jsonschema==3.2.0
    # via drf-spectacular
kombu==5.0.2
    # via celery
prompt-toolkit==3.0.14
    # via click-repl
psycopg2==2.8.6
    # via -r requirements/base.in
pycparser==2.20
    # via cffi
pyjwt==2.0.1
    # via
    #   django-auth-adfs
    #   gemma-zds-client
pyrsistent==0.17.3
    # via jsonschema
python-dateutil==2.8.1
    # via
    #   django-camunda
    #   django-relativedelta
    #   faker
python-decouple==3.4
    # via -r requirements/base.in
python-dotenv==0.15.0
    # via -r requirements/base.in
pytz==2020.5
    # via
    #   -r requirements/base.in
    #   celery
    #   django
pyyaml==5.4.1
    # via
    #   drf-spectacular
    #   gemma-zds-client
redis==3.5.3
    # via django-redis
requests-mock==1.8.0
    # via zgw-consumers
requests==2.25.1
    # via
    #   django-auth-adfs
    #   django-camunda
    #   gemma-zds-client
    #   requests-mock
    #   zgw-consumers
sentry-sdk==0.19.5
    # via -r requirements/base.in
six==1.15.0
    # via
    #   click-repl
    #   cryptography
    #   django-choices
    #   django-compat
    #   jsonschema
    #   python-dateutil
    #   requests-mock
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via faker
uritemplate==3.0.1
    # via drf-spectacular
urllib3==1.26.2
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
uwsgi==2.0.19.1
    # via -r requirements/base.in
vine==5.0.0
    # via
    #   amqp
    #   celery
wcwidth==0.2.5
    # via prompt-toolkit
zgw-consumers==0.15.2
    # via -r requirements/base.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
